refactor: Modularize forms into separate files for improved maintainability
