+++ app/forms/__init__.py
@@ -1,0 +1,10 @@
+from .login_form import LoginForm
+from .registration_form import RegistrationForm
+from .account_form import AccountForm
+from .edit_account_form import EditAccountForm
+from .transaction_form import TransactionForm
+from .category_form import CategoryForm
+from .update_profile_form import UpdateProfileForm
+from .report_form import ReportForm
+from .goal_form import GoalForm
+from .financial_plan_form import FinancialPlanForm
