+++ app/forms/update_profile_form.py
@@ -1,0 +1,28 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, PasswordField, SubmitField
+from wtforms.validators import DataRequired, Email, Length, EqualTo, Optional, ValidationError
+from ..models import User
+
+class UpdateProfileForm(FlaskForm):
+    username = StringField('Nome de Usuário', validators=[DataRequired(), Length(min=4, max=25)])
+    email = StringField('Email', validators=[DataRequired(), Email()])
+    password = PasswordField('Nova Senha', validators=[Optional(), Length(min=6)])
+    confirm_password = PasswordField('Confirmar Nova Senha', validators=[EqualTo('password')])
+    submit = SubmitField('Atualizar')
+
+    def __init__(self, original_username, original_email, *args, **kwargs):
+        super(UpdateProfileForm, self).__init__(*args, **kwargs)
+        self.original_username = original_username
+        self.original_email = original_email
+
+    def validate_username(self, username):
+        if username.data != self.original_username:
+            user = User.query.filter_by(username=self.username.data).first()
+            if user:
+                raise ValidationError('Este nome de usuário já está em uso. Por favor, escolha outro.')
+
+    def validate_email(self, email):
+        if email.data != self.original_email:
+            user = User.query.filter_by(email=self.email.data).first()
+            if user:
+                raise ValidationError('Este email já está registrado. Por favor, use outro.')
